requirements.txt
django==5.1
psycopg2==2.9.9
django-environ==0.11.2
pillow==11.0.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
channels==4.2.0
channels_redis==4.2.1
asgiref==3.8.1